+++ tests/hitl/3_usb_to_can.py
@@ -33,22 +33,6 @@ def test_can_loopback(p):
 
 @test_all_pandas
 @panda_connect_and_init
-def test_safety_nooutput(p):
-  p.set_safety_mode(Panda.SAFETY_SILENT)
-  p.set_can_loopback(True)
-
-  # send a message on bus 0
-  p.can_send(0x1aa, b"message", 0)
-
-  # confirm receive nothing
-  time.sleep(0.05)
-  r = p.can_recv()
-  # bus 192 is messages blocked by TX safety hook on bus 0
-  assert len([x for x in r if x[3] != 192]) == 0
-  assert len([x for x in r if x[3] == 192]) == 1
-
-@test_all_pandas
-@panda_connect_and_init
 def test_reliability(p):
   MSG_COUNT = 100
